feat: 세부 상품 정보 조회 템플릿 & View Controller 구현

+++ src/main/java/gift/controller/CustomErrorController.java
@@ -16,7 +16,7 @@ public class CustomErrorController implements ErrorController {
         Object message = request.getAttribute(RequestDispatcher.ERROR_MESSAGE);
 
         if (status != null && message != null) {
-            model.addAttribute("statusCode", (Integer)status);
+            model.addAttribute("statusCode",status);
             model.addAttribute("errorMessage", message.toString());
         }
         else {
